feat(alloc): support capacity limit
1. Add capacity limit.
2. Add an API to retrieve the current available memory size.

https://github.com/Water-Melon/Melon/issues/32
