+++ src/mln_conf.c
@@ -334,7 +334,7 @@ static inline mln_conf_t *mln_conf_init(void)
     mln_lex_hooks_t hooks;
     struct mln_lex_attr lattr;
 
-    if ((pool = mln_alloc_init(NULL)) == NULL) {
+    if ((pool = mln_alloc_init(NULL, 0)) == NULL) {
         fprintf(stderr, "%s:%d: No memory.\n", __FUNCTION__, __LINE__);
         mln_rbtree_free(cf->domain);
         free(cf);
