+++ src/mln_lang_nil.c
@@ -5,13 +5,6 @@
 #include "mln_lang_nil.h"
 #include <stdio.h>
 
-#ifdef __DEBUG__
-#include <assert.h>
-#define ASSERT(x) assert(x)
-#else
-#define ASSERT(x);
-#endif
-
 static int
 mln_lang_nil_assign(mln_lang_ctx_t *ctx, mln_lang_var_t **ret, mln_lang_var_t *op1, mln_lang_var_t *op2);
 static int
